Move genericPropsChecker to utils folder
