+++ src/cow-react/modules/swap/containers/NewSwapWidget/propsChecker.ts
@@ -3,6 +3,7 @@ import { PriceImpact } from 'hooks/usePriceImpact'
 import { Fraction } from '@uniswap/sdk-core'
 import { ReceiveAmountInfo } from '@cow/modules/swap/helpers/tradeReceiveAmount'
 import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
+import { genericPropsChecker } from '@cow/utils/genericPropsChecker'
 
 function isFractionEqual(prev?: Fraction | null, next?: Fraction | null): boolean {
   return prev && next ? prev.equalTo(next) : prev === next
@@ -57,27 +58,4 @@ export function swapPagePropsChecker(prev: SwapFormProps, next: SwapFormProps): 
     isCurrencyInfoEqual(prev.outputCurrencyInfo, next.outputCurrencyInfo) &&
     isPriceImpactEqual(prev.priceImpactParams, next.priceImpactParams)
   )
-}
-
-export function genericPropsChecker(prev: any, next: any): boolean {
-  if (typeof prev === 'object') {
-    const prevKeys = Object.keys(prev)
-    const nextKeys = Object.keys(next)
-    // Just in case, we take the longest list of keys
-    const keys = prevKeys.length > nextKeys.length ? prevKeys : nextKeys
-
-    return keys.every((key) => {
-      const prevValue = prev[key]
-      const nextValue = next[key]
-
-      // We shouldn't check functions using JSON.stringify because it always returns true
-      if (typeof prevValue === 'function' || typeof nextValue === 'function') {
-        return prevValue === nextValue
-      }
-
-      return JSON.stringify(prevValue) === JSON.stringify(nextValue)
-    })
-  }
-
-  return prev === next
 }
