+++ src/cow-react/modules/swap/pure/TradeRates/index.tsx
@@ -8,8 +8,8 @@ import { Price } from '@cow/modules/swap/pure/Price'
 import TradeGp from 'state/swap/TradeGp'
 import { RowDeadline } from '@cow/modules/swap/containers/Row/RowDeadline'
 import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
-import { genericPropsChecker } from '@cow/modules/swap/containers/NewSwapWidget/propsChecker'
 import { TradeBasicDetails } from '@cow/modules/swap/containers/TradeBasicDetails'
+import { genericPropsChecker } from '@cow/utils/genericPropsChecker'
 
 const SUBSIDY_INFO_MESSAGE_EXTENDED =
   SUBSIDY_INFO_MESSAGE + '. Click on the discount button on the right for more info.'
